no threads, only pitch ladder
