+++ app/prj.conf
@@ -25,8 +25,9 @@ CONFIG_REBOOT=y
 # CONFIG_CONSOLE=y
 # CONFIG_UART_CONSOLE=y
 CONFIG_DEBUG=y
-CONFIG_MAIN_STACK_SIZE=16384
-CONFIG_SYSTEM_WORKQUEUE_STACK_SIZE=16384
+#16384
+CONFIG_MAIN_STACK_SIZE=8192
+CONFIG_SYSTEM_WORKQUEUE_STACK_SIZE=8192
 
 #CONFIG_SENSOR=y
 CONFIG_BLINK=y
@@ -57,7 +58,7 @@ CONFIG_LVGL=y
 
 CONFIG_LV_MEM_CUSTOM=y
 #2048 default 8192 16384
-CONFIG_LV_Z_MEM_POOL_SIZE=8192
+CONFIG_LV_Z_MEM_POOL_SIZE=2048
 
 CONFIG_LV_USE_LOG=y
 CONFIG_LV_USE_LABEL=y
@@ -76,6 +77,7 @@ CONFIG_LV_USE_CANVAS=y
 #CONFIG_LV_MEM_SIZE=(48U * 1024U)
 #CONFIG_LV_USE_LOG=y
 #CONFIG_LV_LOG_LEVEL_INFO=y
+#CONFIG_LV_USE_OS
 
 CONFIG_LV_COMPASS=y
 #CONFIG_LV_COMPASS_GET_VALUE_DEFAULT=44
