+++ app/src/main.c
@@ -19,7 +19,7 @@
 
 #include <app/drivers/blink.h>
 // #include <app/lib/lv_hud.h>
-#include <app/lib/lv_compass.h>
+//#include <app/lib/lv_compass.h>
 #include <app/lib/lv_pitch_ladder.h>
 
 #include <app_version.h>
@@ -30,7 +30,7 @@ LOG_MODULE_REGISTER(main, CONFIG_APP_LOG_LEVEL);
 #define BLINK_PERIOD_MS_STEP 100U
 #define BLINK_PERIOD_MS_MAX  1300U
 
-#define TICK_PERIOD   (2000)
+#define TICK_PERIOD   (1000)
 
 #define LED0_NODE	DT_ALIAS(led0)
 static const struct gpio_dt_spec led = GPIO_DT_SPEC_GET(LED0_NODE, gpios);
@@ -85,8 +85,8 @@ void pitch_ladder_update(int screen_id, int param_id);
 void lv_pitch_ladder_init(lv_style_t *indicator_style);
 void hud_style(lv_style_t *indicator_style);
 
-K_TIMER_DEFINE(control_timer, display_timer_handler, NULL);
-K_WORK_DEFINE(control_work, control_task_handler);
+//K_TIMER_DEFINE(control_timer, display_timer_handler, NULL);
+//K_WORK_DEFINE(control_work, control_task_handler);
 
 void numeric_param_update(int screen_id, int param_id, const char *format)
 {
@@ -119,48 +119,48 @@ void  hud_style(lv_style_t *indicator_style)
 /**
  * A simple horizontal comapss
  */
-void lv_example_comapss_1(lv_style_t *indicator_style)
-{
-    //lv_obj_t * comapss = lv_comapss_create(lv_scr_act());
-	compass_obj = lv_comapss_create(lv_scr_act());
-    lv_obj_set_size(compass_obj, 80, 110);
-    lv_comapss_set_mode(compass_obj, LV_COMAPSS_MODE_HORIZONTAL_BOTTOM);
+// void lv_example_comapss_1(lv_style_t *indicator_style)
+// {
+//     //lv_obj_t * comapss = lv_comapss_create(lv_scr_act());
+// 	compass_obj = lv_comapss_create(lv_scr_act());
+//     lv_obj_set_size(compass_obj, 80, 110);
+//     lv_comapss_set_mode(compass_obj, LV_COMAPSS_MODE_HORIZONTAL_BOTTOM);
 
-    /*static lv_style_t indicator_style;
-    lv_style_init(&indicator_style);
-    // Label style properties 
-    lv_style_set_text_font(&indicator_style, LV_FONT_DEFAULT);
-    lv_style_set_text_color(&indicator_style, lv_color_black());	//lv_color_hex(0x0000ff)
-    // Major tick properties 
-    lv_style_set_line_color(&indicator_style, lv_color_black());
-    lv_style_set_width(&indicator_style, 20U); // Tick length
-    lv_style_set_line_width(&indicator_style, 2U); // Tick width
-	*/
-    lv_obj_add_style(compass_obj, indicator_style, LV_PART_INDICATOR);
+//     // static lv_style_t indicator_style;
+//     // lv_style_init(&indicator_style);
+//     // // Label style properties 
+//     // lv_style_set_text_font(&indicator_style, LV_FONT_DEFAULT);
+//     // lv_style_set_text_color(&indicator_style, lv_color_black());	//lv_color_hex(0x0000ff)
+//     // // Major tick properties 
+//     // lv_style_set_line_color(&indicator_style, lv_color_black());
+//     // lv_style_set_width(&indicator_style, 20U); // Tick length
+//     // lv_style_set_line_width(&indicator_style, 2U); // Tick width
+	
+//     lv_obj_add_style(compass_obj, indicator_style, LV_PART_INDICATOR);
 
-	lv_obj_add_style(compass_obj, indicator_style, LV_PART_ITEMS);
-	lv_obj_add_style(compass_obj, indicator_style, LV_PART_MAIN);
+// 	lv_obj_add_style(compass_obj, indicator_style, LV_PART_ITEMS);
+// 	lv_obj_add_style(compass_obj, indicator_style, LV_PART_MAIN);
 
-    lv_obj_center(compass_obj);
+//     lv_obj_center(compass_obj);
 
-    lv_comapss_set_label_show(compass_obj, true);
+//     lv_comapss_set_label_show(compass_obj, true);
 
-    lv_comapss_set_total_tick_count(compass_obj, 5);//11
-    lv_comapss_set_major_tick_every(compass_obj, 2);//5
+//     lv_comapss_set_total_tick_count(compass_obj, 5);//11
+//     lv_comapss_set_major_tick_every(compass_obj, 2);//5
 
-	lv_comapss_set_direction_angle(compass_obj, 45);
+// 	lv_comapss_set_direction_angle(compass_obj, 45);
 
-	//lv_obj_align_to(compass_obj, screens[0].screen, LV_ALIGN_BOTTOM_MID, 0, 0);
+// 	//lv_obj_align_to(compass_obj, screens[0].screen, LV_ALIGN_BOTTOM_MID, 0, 0);
 
-	//lv_obj_set_align(compass_obj, LV_ALIGN_CENTER);//LV_ALIGN_TOP_MID LV_ALIGN_CENTER
-	//lv_obj_set_pos(compass_obj, 0, 0);
+// 	//lv_obj_set_align(compass_obj, LV_ALIGN_CENTER);//LV_ALIGN_TOP_MID LV_ALIGN_CENTER
+// 	//lv_obj_set_pos(compass_obj, 0, 0);
 
-	// pitch_ladder_obj = lv_comapss_create(lv_scr_act());
-    // lv_obj_set_size(pitch_ladder_obj, 80, 110);
-    // lv_comapss_set_mode(pitch_ladder_obj, LV_COMAPSS_MODE_HORIZONTAL_BOTTOM);
-}
+// 	// pitch_ladder_obj = lv_comapss_create(lv_scr_act());
+//     // lv_obj_set_size(pitch_ladder_obj, 80, 110);
+//     // lv_comapss_set_mode(pitch_ladder_obj, LV_COMAPSS_MODE_HORIZONTAL_BOTTOM);
+// }
 
-void lv_pitch_ladder_init(lv_style_t *indicator_style)
+/*void lv_pitch_ladder_init(lv_style_t *indicator_style)
 {
 	pitch_ladder_obj = lv_pitch_ladder_create(lv_scr_act());
     lv_obj_set_size(pitch_ladder_obj, 120, 120);
@@ -181,14 +181,14 @@ void lv_pitch_ladder_init(lv_style_t *indicator_style)
 	lv_obj_align_to(pitch_ladder_obj, screens[0].screen, LV_ALIGN_CENTER, 0, 0);
 
 	//lv_pitch_ladder_set_direction_angle(pitch_ladder_obj, 45);
-	lv_pitch_ladder_set_angle(pitch_ladder_obj, 0);
+	lv_pitch_ladder_set_angle(pitch_ladder_obj, 4);
 
-}
+}*/
 
-void display_timer_handler(struct k_timer * timer)
+/*void display_timer_handler(struct k_timer * timer)
 {
     k_work_submit(&control_work);
-}
+}*/
 
 void control_task_handler(struct k_work * work)
 {  
@@ -215,15 +215,18 @@ void control_task_handler(struct k_work * work)
 	screen0_x_value = gyr[1].val1;
 	screen0_y_value = gyr[2].val1;
 	//screen0_y_value = gyr[2].val1;//custom_get_value(0);
-    numeric_param_update(screen_id, 0, "h: %d");
+    /*numeric_param_update(screen_id, 0, "h: %d");
 	numeric_param_update(screen_id, 1, "p: %d");
-	numeric_param_update(screen_id, 2, "r: %d");
+	numeric_param_update(screen_id, 2, "r: %d");*/
 	compass_value = gyr[0].val1;
 	//hud_update(screen_id, 3);
 	pitch_ladder_pitch = gyr[0].val1*10;
-	pitch_ladder_update(screen_id, 4);
-	lv_task_handler();
+	//pitch_ladder_update(screen_id, 4);
+	//lv_task_handler();
+	//printk("app event handler\n");
 }
+lv_obj_t *canvas;
+lv_color_t buffer[LV_CANVAS_BUF_SIZE_TRUE_COLOR_ALPHA(32, 18)];
 
 void display_screens_init(void)
 {
@@ -247,15 +250,15 @@ void display_screens_init(void)
 	lv_scr_load(screens[0].screen);
 
 	static lv_style_t indicator_style;
-	hud_style(&indicator_style);
+	//hud_style(&indicator_style);
 	//lv_example_comapss_1(&indicator_style);
-	lv_pitch_ladder_init(&indicator_style);
+	//lv_pitch_ladder_init(&indicator_style);
     LOG_INF("main init");
 	// lv_obj_t *hello_world_label = lv_label_create(lv_scr_act());
 	// lv_label_set_text(hello_world_label, "WIKTOR");
 	// lv_obj_align_to(hello_world_label, screens[0].screen, LV_ALIGN_TOP_RIGHT, 0, 0);
 	
-
+/*
 	screen0_label_obj = lv_label_create(lv_scr_act());
 	lv_label_set_text(screen0_label_obj, "11");
 	lv_obj_align_to(screen0_label_obj, NULL, LV_ALIGN_BOTTOM_LEFT, 0, 0);
@@ -270,12 +273,32 @@ void display_screens_init(void)
 	lv_label_set_text(screen0_y_obj, "33");
 	lv_obj_align_to(screen0_y_obj, NULL, LV_ALIGN_BOTTOM_RIGHT, -15, 0);
 	screen0_y_value = 33;
+*/
+pitch_ladder_obj = lv_pitch_ladder_create(lv_scr_act());
 
+//canvas = lv_canvas_create(lv_scr_act());
+//lv_canvas_set_buffer(canvas, buffer,32, 18, LV_IMG_CF_TRUE_COLOR_ALPHA);
 }
 
+/*static void lv_pitch_ladder_draw_numeric( int16_t x, int16_t y, int16_t angle)
+{
+    //lv_obj_t *canvas_num = lv_canvas_create(obj);
+
+    
+    lv_img_set_angle(canvas, angle);
+    lv_obj_align(canvas, LV_ALIGN_CENTER, x, y);
+
+    lv_color_t c = lv_color_make(0xFF, 0xFF, 0xFF);
+    lv_canvas_fill_bg(canvas, c, LV_OPA_COVER); 
+
+    lv_draw_label_dsc_t lbl_dsc;
+    lv_draw_label_dsc_init(&lbl_dsc);
+
+    lv_canvas_draw_text(canvas, 2, 2, 30, &lbl_dsc, "-90");
+}*/
 
 
-void hud_update(int screen_id, int param_id)
+/*void hud_update(int screen_id, int param_id)
 {
     param_t * param = &screens[screen_id].params[param_id];
     if (param == NULL || *param->object == NULL)
@@ -285,7 +308,7 @@ void hud_update(int screen_id, int param_id)
 		lv_comapss_set_direction_angle(*param->object, *param->value);
         return;
     }
-}
+}*/
 
 void pitch_ladder_update(int screen_id, int param_id)
 {
@@ -297,7 +320,7 @@ void pitch_ladder_update(int screen_id, int param_id)
 		//lv_pitch_ladder_set_pitch(*param->object, *param->value);
 		//lv_pitch_ladder_set_roll(*param->object, *param->value);
 
-		lv_pitch_ladder_set_angle(*param->object, *param->value);
+		//lv_pitch_ladder_set_angle(*param->object, *param->value);
         return;
     }
 }
@@ -339,10 +362,10 @@ int main(void)
     }
 	//LOG_ERR("Error test");
 	//LOG_INF("inf test");
-	if (!device_is_ready(gyro_dev)) {
+	/*if (!device_is_ready(gyro_dev)) {
 		printk("Device %s is not ready\n", gyro_dev->name);
 		return 0;
-	}
+	}*/
 
 	printk("Device initialization success\n");
     display_screens_init();
@@ -351,7 +374,7 @@ int main(void)
 	display_blanking_off(display_dev);
 	
 	blink_set_period_ms(blink, period_ms);
-	k_timer_start(&control_timer, K_MSEC(TICK_PERIOD), K_MSEC(TICK_PERIOD));
+	/*k_timer_start(&control_timer, K_MSEC(TICK_PERIOD), K_MSEC(TICK_PERIOD));
 
 	
 
@@ -361,11 +384,18 @@ int main(void)
 	printk("GX: %d.%d; GY: %d.%d; GZ: %d.%d;\n",
 			gyr[0].val1, gyr[0].val2,
 			gyr[1].val1, gyr[1].val2,
-			gyr[2].val1, gyr[2].val2);
-
+			gyr[2].val1, gyr[2].val2);*/
+	unsigned int angle = 0;
 	while (1) {
-
-        k_msleep(SLEEP_TIME_MS);
+		//lv_pitch_ladder_draw_numeric(0,0,angle);
+		if(angle<3600) angle += 100; else angle = 0;
+		//lv_refr_now(NULL);
+		lv_task_handler();
+		//lv_lock();
+		
+		//lv_unlock();
+        k_msleep(1000);//SLEEP_TIME_MS);
+		lv_pitch_ladder_set_pitch(pitch_ladder_obj, angle);
 	}
 
 	return 0;
