+++ include/app/lib/lv_pitch_ladder.h
@@ -22,12 +22,6 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
  */
 
-// #ifndef A3688119_289C_497B_9268_272425E79E01
-// #define A3688119_289C_497B_9268_272425E79E01
-
-// #ifndef B70830EF_BB65_4CEE_A11C_5BFFAC319948
-// #define B70830EF_BB65_4CEE_A11C_5BFFAC319948
-
 #ifndef APP_LIB_LV_PITCH_LADDER_H_
 #define APP_LIB_LV_PITCH_LADDER_H_
 
@@ -68,14 +62,16 @@ LV_EXPORT_CONST_INT(LV_PITCH_LADDER_MAJOR_TICK_EVERY_DEFAULT);
 #define LV_PITCH_LADDER_LABEL_ENABLED_DEFAULT (1U)
 LV_EXPORT_CONST_INT(LV_PITCH_LADDER_LABEL_ENABLED_DEFAULT);
 
+#define LV_PITCH_LADDE_CANVAS_WIDTH 80
+#define LV_PITCH_LADDE_CANVAS_HEIGHT 48
+#define LV_PITCH_LADDER_HORIZ_GAP    LV_PITCH_LADDE_CANVAS_WIDTH/4
+#define LV_PITCH_LADDER_HORIZ_LEND (LV_PITCH_LADDE_CANVAS_WIDTH -LV_PITCH_LADDER_HORIZ_GAP)/2
+#define LV_PITCH_LADDER_HORIZ_RSTART LV_PITCH_LADDER_HORIZ_LEND + LV_PITCH_LADDER_HORIZ_GAP
+#define LV_PITCH_LADDER_AIM_W 8
 /**********************
  *      TYPEDEFS
  **********************/
-// enum _lv_result_t {
-//     LV_RESULT_INVALID = 0, /*Typically indicates that the object is deleted (become invalid) in the action
-//                       function or an operation was failed*/
-//     LV_RESULT_OK,      /*The object is valid (no deleted) after the action*/
-// };
+
 /**
  * Pitch_ladder mode
  */
@@ -91,19 +87,8 @@ enum {
 typedef uint32_t lv_pitch_ladder_mode_t;
 
 typedef struct {
-    lv_style_t * main_style;
     lv_style_t * indicator_style;
     lv_style_t * items_style;
-    int32_t minor_range;
-    int32_t major_range;
-    uint32_t first_tick_idx_in_section;
-    uint32_t last_tick_idx_in_section;
-    uint32_t first_tick_idx_is_major;
-    uint32_t last_tick_idx_is_major;
-    int32_t first_tick_in_section_width;
-    int32_t last_tick_in_section_width;
-    lv_point_t first_tick_in_section;
-    lv_point_t last_tick_in_section;
 } lv_pitch_ladder_section_t;
 
 typedef struct {
@@ -111,23 +96,14 @@ typedef struct {
     lv_ll_t section_ll;     /**< Linked list for the sections (stores lv_pitch_ladder_section_t)*/
     const char ** txt_src;
     lv_pitch_ladder_mode_t mode;
-    int32_t range_min;
-    int32_t range_max;
     int32_t dir_angle;
     int32_t pitch_angle;
     int32_t roll_angle;
-    uint32_t total_tick_count   : 15;
-    uint32_t major_tick_every   : 15;
-    uint32_t label_enabled      : 1;
     uint32_t post_draw          : 1;
-    uint32_t draw_ticks_on_top  : 1;
-    /* Round pitch_ladder */
-    uint32_t angle_range;
-    int32_t rotation;
+    uint32_t widget_draw          : 1;
+    //lv_obj_t *canvas;
+    //lv_color_t buffer[3*LV_PITCH_LADDE_CANVAS_WIDTH*LV_PITCH_LADDE_CANVAS_HEIGHT];//[3*32*18];
     /* Private properties */
-    int32_t custom_label_cnt;
-    int32_t last_tick_width;
-    int32_t first_tick_width;
 } lv_pitch_ladder_t;
 
 extern const lv_obj_class_t lv_pitch_ladder_class;
@@ -150,50 +126,6 @@ lv_obj_t * lv_pitch_ladder_create(lv_obj_t * parent);
 /*=====================
  * Setter functions
  *====================*/
-
-/**
- * Set pitch_ladder mode. 
- * @param obj       pointer the pitch_ladder object
- * @param mode      the new pitch_ladder mode
- */
-void lv_pitch_ladder_set_mode(lv_obj_t * obj, lv_pitch_ladder_mode_t mode);
-
-/**
- * Set pitch_ladder total tick count (including minor and major ticks)
- * @param obj       pointer the pitch_ladder object
- * @param total_tick_count    New total tick count
- */
-void lv_pitch_ladder_set_total_tick_count(lv_obj_t * obj, uint32_t total_tick_count);
-
-/**
- * Sets how often the major tick will be drawn
- * @param obj                 pointer the pitch_ladder object
- * @param major_tick_every    the new count for major tick drawing
- */
-void lv_pitch_ladder_set_major_tick_every(lv_obj_t * obj, uint32_t major_tick_every);
-
-/**
- * Sets label visibility
- * @param obj           pointer the pitch_ladder object
- * @param show_label    true/false to enable tick label
- */
-void lv_pitch_ladder_set_label_show(lv_obj_t * obj, bool show_label);
-
-/**
- * Set the minimal and maximal values on a pitch_ladder
- * @param obj       pointer to a pitch_ladder object
- * @param min       minimum value of the pitch_ladder
- * @param max       maximum value of the pitch_ladder
- */
-void lv_pitch_ladder_set_range(lv_obj_t * obj, int32_t min, int32_t max);
-
-/**
- * Set the direction angle value on a pitch_ladder
- * @param obj           pointer to a pitch_ladder object
- * @param direction       direction angle of the pitch_ladder
- */
-void lv_pitch_ladder_set_direction_angle(lv_obj_t * obj, int32_t direction);
-
 /**
  * Set the pitch value on a pitch_ladder
  * @param obj           pointer to a pitch_ladder object
@@ -201,169 +133,10 @@ void lv_pitch_ladder_set_direction_angle(lv_obj_t * obj, int32_t direction);
  */
 void lv_pitch_ladder_set_pitch(lv_obj_t * obj, int32_t pitch);
 
-/**
- * Set the roll value on a pitch_ladder
- * @param obj           pointer to a pitch_ladder object
- * @param roll       roll angle of the pitch_ladder
- */
-void lv_pitch_ladder_set_roll(lv_obj_t * obj, int32_t roll);
-
-/**
- * Limit the tick values to 0 .. 360
- * @param value           pointer to a pitch_ladder object
- */
-int32_t lv_pitch_ladder_limit(int32_t value);
-
-/**
- * Set properties specific to round pitch_ladder
- * @param obj           pointer to a pitch_ladder object
- * @param angle_range   the angular range of the pitch_ladder
- */
-void lv_pitch_ladder_set_angle_range(lv_obj_t * obj, uint32_t angle_range);
-
-/**
- * Set properties specific to round pitch_ladder
- * @param obj       pointer to a pitch_ladder object
- * @param rotation  the angular offset from the 3 o'clock position (clock-wise)
- */
-void lv_pitch_ladder_set_rotation(lv_obj_t * obj, int32_t rotation);
-
-/**
- * Point the needle to the corresponding value through the line
- * @param obj              pointer to a pitch_ladder object
- * @param needle_line      needle_line of the pitch_ladder. The line points will be allocated and
- *                         managed by the pitch_ladder unless the line point array was previously set
- *                         using `lv_line_set_points_mutable`.
- * @param needle_length    length of the needle
- *                         needle_length>0 needle_length=needle_length;
- *                         needle_length<0 needle_length=radius-|needle_length|;
- * @param value            needle to point to the corresponding value
- */
-void lv_pitch_ladder_set_line_needle_value(lv_obj_t * obj, lv_obj_t * needle_line, int32_t needle_length,
-                                    int32_t value);
-
-/**
- * Point the needle to the corresponding value through the image,
-   image must point to the right. E.g. -O------>
- * @param obj              pointer to a pitch_ladder object
- * @param needle_img       needle_img of the pitch_ladder
- * @param value            needle to point to the corresponding value
- */
-void lv_pitch_ladder_set_image_needle_value(lv_obj_t * obj, lv_obj_t * needle_img, int32_t value);
-
-/**
- * Set custom text source for major ticks labels
- * @param obj       pointer to a pitch_ladder object
- * @param txt_src   pointer to an array of strings which will be display at major ticks
- */
-void lv_pitch_ladder_set_text_src(lv_obj_t * obj, const char * txt_src[]);
-
-/**
- * Draw the pitch_ladder after all the children are drawn
- * @param obj       pointer to a pitch_ladder object
- * @param en        true: enable post draw
- */
-void lv_pitch_ladder_set_post_draw(lv_obj_t * obj, bool en);
-
-/**
- * Draw the pitch_ladder ticks on top of all parts
- * @param obj       pointer to a pitch_ladder object
- * @param en        true: enable draw ticks on top of all parts
- */
-void lv_pitch_ladder_set_draw_ticks_on_top(lv_obj_t * obj, bool en);
-
-/**
- * Add a section to the given pitch_ladder
- * @param obj       pointer to a pitch_ladder object
- * @return          pointer to the new section
- */
-lv_pitch_ladder_section_t * lv_pitch_ladder_add_section(lv_obj_t * obj);
-
-/**
- * Set the range for the given pitch_ladder section
- * @param section       pointer to a pitch_ladder section object
- * @param minor_range   section new minor range
- * @param major_range   section new major range
- */
-void lv_pitch_ladder_section_set_range(lv_pitch_ladder_section_t * section, int32_t minor_range, int32_t major_range);
-
-/**
- * Set the style of the part for the given pitch_ladder section
- * @param section   pointer to a pitch_ladder section object
- * @param part      the part for the section, e.g. LV_PART_INDICATOR
- * @param section_part_style Pointer to the section part style
- */
-void lv_pitch_ladder_section_set_style(lv_pitch_ladder_section_t * section, lv_part_t part, lv_style_t * section_part_style);
-
 /*=====================
  * Getter functions
  *====================*/
 
-/**
- * Get pitch_ladder mode.
- * @param obj   pointer the pitch_ladder object
- * @return      Pitch_ladder mode
- */
-lv_pitch_ladder_mode_t lv_pitch_ladder_get_mode(lv_obj_t * obj);
-
-/**
- * Get pitch_ladder total tick count (including minor and major ticks)
- * @param obj   pointer the pitch_ladder object
- * @return      Pitch_ladder total tick count
- */
-int32_t lv_pitch_ladder_get_total_tick_count(lv_obj_t * obj);
-
-/**
- * Gets how often the major tick will be drawn
- * @param obj   pointer the pitch_ladder object
- * @return      Pitch_ladder major tick every count
- */
-int32_t lv_pitch_ladder_get_major_tick_every(lv_obj_t * obj);
-
-/**
- * Gets label visibility
- * @param obj   pointer the pitch_ladder object
- * @return      true if tick label is enabled, false otherwise
- */
-bool lv_pitch_ladder_get_label_show(lv_obj_t * obj);
-
-/**
- * Get angle range of a round pitch_ladder
- * @param obj   pointer to a pitch_ladder object
- * @return      Pitch_ladder angle_range
- */
-uint32_t lv_pitch_ladder_get_angle_range(lv_obj_t * obj);
-
-/**
- * Get the min range for the given pitch_ladder section
- * @param obj   pointer to a pitch_ladder section object
- * @return      section minor range
- */
-int32_t lv_pitch_ladder_get_range_min_value(lv_obj_t * obj);
-
-/**
- * Get the max range for the given pitch_ladder section
- * @param obj   pointer to a pitch_ladder section object
- * @return      section max range
- */
-int32_t lv_pitch_ladder_get_range_max_value(lv_obj_t * obj);
-
-
-void lv_pitch_ladder_set_angle(lv_obj_t * obj, int16_t rotation);
-/**********************
- *      MACROS
- **********************/
-
-//#endif /*LV_USE_PITCH_LADDER*/
-
-// #ifdef __cplusplus
-// } /*extern "C"*/
-// #endif
 
 #endif /*APP_LIB_LV_PITCH_LADDER_H*/
 
-
-// #endif /* B70830EF_BB65_4CEE_A11C_5BFFAC319948 */
-
-
-// #endif /* A3688119_289C_497B_9268_272425E79E01 */
